+++ src/main/java/org/jenkinsci/plugins/fodupload/SharedCreateApplicationForm.java
@@ -2,6 +2,7 @@ package org.jenkinsci.plugins.fodupload;
 
 import net.sf.json.JSONNull;
 import net.sf.json.JSONObject;
+import org.jenkinsci.plugins.fodupload.FodApi.FodApiConnection;
 import org.jenkinsci.plugins.fodupload.controllers.ApplicationsController;
 import org.jenkinsci.plugins.fodupload.controllers.AttributesController;
 import org.jenkinsci.plugins.fodupload.models.*;
@@ -18,7 +19,7 @@ public class SharedCreateApplicationForm {
     //<editor-fold desc="Create Application">
 
     public static Result<CreateApplicationResponse> submitCreateApplication(AuthenticationModel authModel, JSONObject formObject) throws IOException {
-        FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+        FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
         String correlationId = Utils.createCorrelationId();
 
         System.out.println("Creating an application with FOD API. [CorrelationId = " + correlationId + "]");
@@ -95,7 +96,7 @@ public class SharedCreateApplicationForm {
     //<editor-fold desc="Create Microservice">
 
     public static Result<Integer> submitCreateMicroservice(AuthenticationModel authModel, JSONObject formObject) throws IOException {
-        FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+        FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
         String correlationId = Utils.createCorrelationId();
 
         System.out.println("Creating a microservice with FOD API. [CorrelationId = " + correlationId + "]");
@@ -142,7 +143,7 @@ public class SharedCreateApplicationForm {
     //<editor-fold desc="Create Release">
 
     public static Result<Integer> submitCreateRelease(AuthenticationModel authModel, JSONObject formObject) throws IOException {
-        FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+        FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
         String correlationId = Utils.createCorrelationId();
 
         System.out.println("Creating a release with FOD API. [CorrelationId = " + correlationId + "]");
