+++ src/main/java/org/jenkinsci/plugins/fodupload/StaticAssessmentBuildStep.java
@@ -14,6 +14,7 @@ import hudson.util.ListBoxModel;
 import jenkins.model.Jenkins;
 import jenkins.tasks.SimpleBuildStep;
 import net.sf.json.JSONObject;
+import org.jenkinsci.plugins.fodupload.FodApi.FodApiConnection;
 import org.jenkinsci.plugins.fodupload.actions.CrossBuildAction;
 import org.jenkinsci.plugins.fodupload.controllers.*;
 import org.jenkinsci.plugins.fodupload.models.AuthenticationModel;
@@ -35,6 +36,9 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+/*
+ Freestyle Build Step
+ */
 
 @SuppressWarnings("unused")
 public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildStep {
@@ -186,7 +190,7 @@ public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildSt
             authModel = AuthenticationModel.fromPersonalAccessToken(username, personalAccessToken, tenantId);
         }
 
-        return ApiConnectionFactory.createApiConnection(authModel);
+        return ApiConnectionFactory.createApiConnection(authModel, false, null);
     }
 
     private boolean isTechStackWithLanguageLevel(int techStack) {
@@ -570,7 +574,7 @@ public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildSt
         public String retrieveCurrentUserSession(JSONObject authModelObject) {
             try {
                 AuthenticationModel authModel = Utils.getAuthModelFromObject(authModelObject);
-                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
                 UsersController usersController = new UsersController(apiConnection, null, Utils.createCorrelationId());
 
                 return Utils.createResponseViewModel(usersController.getCurrentUserSession());
@@ -584,7 +588,7 @@ public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildSt
         public String retrieveAssessmentTypeEntitlements(Integer releaseId, JSONObject authModelObject) {
             try {
                 AuthenticationModel authModel = Utils.getAuthModelFromObject(authModelObject);
-                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
                 AssessmentTypesController assessmentTypesController = new AssessmentTypesController(apiConnection, null, Utils.createCorrelationId());
 
                 return Utils.createResponseViewModel(assessmentTypesController.getStaticAssessmentTypeEntitlements(releaseId));
@@ -598,7 +602,7 @@ public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildSt
         public String retrieveAuditPreferences(Integer releaseId, Integer assessmentType, Integer frequencyType, JSONObject authModelObject) {
             try {
                 AuthenticationModel authModel = Utils.getAuthModelFromObject(authModelObject);
-                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
                 ReleaseController releaseController = new ReleaseController(apiConnection, null, Utils.createCorrelationId());
 
                 return Utils.createResponseViewModel(releaseController.getAuditPreferences(releaseId, assessmentType, frequencyType));
@@ -612,7 +616,7 @@ public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildSt
         public String retrieveLookupItems(String type, JSONObject authModelObject) {
             try {
                 AuthenticationModel authModel = Utils.getAuthModelFromObject(authModelObject);
-                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
                 LookupItemsController lookupItemsController = new LookupItemsController(apiConnection, null, Utils.createCorrelationId());
 
                 return Utils.createResponseViewModel(lookupItemsController.getLookupItems(FodEnums.APILookupItemTypes.valueOf(type)));
@@ -626,7 +630,7 @@ public class StaticAssessmentBuildStep extends Recorder implements SimpleBuildSt
         public String retrieveStaticScanSettings(Integer releaseId, JSONObject authModelObject) {
             try {
                 AuthenticationModel authModel = Utils.getAuthModelFromObject(authModelObject);
-                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel);
+                FodApiConnection apiConnection = ApiConnectionFactory.createApiConnection(authModel, false, null);
                 StaticScanController staticScanController = new StaticScanController(apiConnection, null, Utils.createCorrelationId());
 
                 return Utils.createResponseViewModel(staticScanController.getStaticScanSettings(releaseId));
