+++ src/main/java/org/jenkinsci/plugins/fodupload/controllers/ApplicationsController.java
@@ -2,7 +2,7 @@ package org.jenkinsci.plugins.fodupload.controllers;
 
 import com.google.gson.reflect.TypeToken;
 import okhttp3.*;
-import org.jenkinsci.plugins.fodupload.FodApiConnection;
+import org.jenkinsci.plugins.fodupload.FodApi.FodApiConnection;
 import org.jenkinsci.plugins.fodupload.Json;
 import org.jenkinsci.plugins.fodupload.Utils;
 import org.jenkinsci.plugins.fodupload.models.CreateApplicationModel;
