+++ src/main/java/org/jenkinsci/plugins/fodupload/controllers/UsersController.java
@@ -4,7 +4,7 @@ import com.google.gson.reflect.TypeToken;
 import okhttp3.HttpUrl;
 import okhttp3.Request;
 import okhttp3.Response;
-import org.jenkinsci.plugins.fodupload.FodApiConnection;
+import org.jenkinsci.plugins.fodupload.FodApi.FodApiConnection;
 import org.jenkinsci.plugins.fodupload.models.Result;
 import org.jenkinsci.plugins.fodupload.models.response.CurrentUserSessionResponse;
